+++ lib/index.ts
@@ -3,7 +3,9 @@
  * The software in this package is published under the terms of the MIT license,
  * a copy of which has been included with this distribution in the LICENSE.txt file.
  */
-export * from './SoqlParser';
-export * from './SoqlComposer';
-export { formatQuery, FormatOptions } from './SoqlFormatter';
+export * from './publicUtils';
 export * from './models/SoqlQuery.model';
+export { ComposeField } from './publicUtils';
+export * from './SoqlComposer';
+export { FormatOptions } from './SoqlFormatter';
+export * from './SoqlParser';